\name{addMGSM27}
\alias{addMGSM27}
\docType{data}
\title{
Data for 5 additional cell types present in multiple myeloma whole bone marrow
}
\description{
Osteoclasts, Adipocytes, Osteoblasts, Multiple Myeloma Plasma Cells, and Plasma Memory Cells
}
\usage{data("addMGSM27")}
\format{
  A data frame with 25783 observations on the following 100 variables.
  \describe{
    \item{\code{osteoclast}}{a numeric vector}
    \item{\code{osteoclast.1}}{a numeric vector}
    \item{\code{osteoclast.2}}{a numeric vector}
    \item{\code{osteoclast.3}}{a numeric vector}
    \item{\code{osteoclast.4}}{a numeric vector}
    \item{\code{osteoclast.5}}{a numeric vector}
    \item{\code{osteoclast.6}}{a numeric vector}
    \item{\code{osteoclast.7}}{a numeric vector}
    \item{\code{osteoclast.8}}{a numeric vector}
    \item{\code{osteoclast.9}}{a numeric vector}
    \item{\code{MM.plasma.cell}}{a numeric vector}
    \item{\code{MM.plasma.cell.1}}{a numeric vector}
    \item{\code{MM.plasma.cell.2}}{a numeric vector}
    \item{\code{MM.plasma.cell.3}}{a numeric vector}
    \item{\code{MM.plasma.cell.4}}{a numeric vector}
    \item{\code{MM.plasma.cell.5}}{a numeric vector}
    \item{\code{MM.plasma.cell.6}}{a numeric vector}
    \item{\code{MM.plasma.cell.7}}{a numeric vector}
    \item{\code{MM.plasma.cell.8}}{a numeric vector}
    \item{\code{MM.plasma.cell.9}}{a numeric vector}
    \item{\code{MM.plasma.cell.10}}{a numeric vector}
    \item{\code{MM.plasma.cell.11}}{a numeric vector}
    \item{\code{MM.plasma.cell.12}}{a numeric vector}
    \item{\code{MM.plasma.cell.13}}{a numeric vector}
    \item{\code{MM.plasma.cell.14}}{a numeric vector}
    \item{\code{MM.plasma.cell.15}}{a numeric vector}
    \item{\code{MM.plasma.cell.16}}{a numeric vector}
    \item{\code{MM.plasma.cell.17}}{a numeric vector}
    \item{\code{MM.plasma.cell.18}}{a numeric vector}
    \item{\code{MM.plasma.cell.19}}{a numeric vector}
    \item{\code{MM.plasma.cell.20}}{a numeric vector}
    \item{\code{MM.plasma.cell.21}}{a numeric vector}
    \item{\code{MM.plasma.cell.22}}{a numeric vector}
    \item{\code{MM.plasma.cell.23}}{a numeric vector}
    \item{\code{MM.plasma.cell.24}}{a numeric vector}
    \item{\code{MM.plasma.cell.25}}{a numeric vector}
    \item{\code{MM.plasma.cell.26}}{a numeric vector}
    \item{\code{MM.plasma.cell.27}}{a numeric vector}
    \item{\code{MM.plasma.cell.28}}{a numeric vector}
    \item{\code{MM.plasma.cell.29}}{a numeric vector}
    \item{\code{MM.plasma.cell.30}}{a numeric vector}
    \item{\code{MM.plasma.cell.31}}{a numeric vector}
    \item{\code{MM.plasma.cell.32}}{a numeric vector}
    \item{\code{MM.plasma.cell.33}}{a numeric vector}
    \item{\code{MM.plasma.cell.34}}{a numeric vector}
    \item{\code{MM.plasma.cell.35}}{a numeric vector}
    \item{\code{MM.plasma.cell.36}}{a numeric vector}
    \item{\code{MM.plasma.cell.37}}{a numeric vector}
    \item{\code{MM.plasma.cell.38}}{a numeric vector}
    \item{\code{MM.plasma.cell.39}}{a numeric vector}
    \item{\code{MM.plasma.cell.40}}{a numeric vector}
    \item{\code{MM.plasma.cell.41}}{a numeric vector}
    \item{\code{osteoblast}}{a numeric vector}
    \item{\code{osteoblast.1}}{a numeric vector}
    \item{\code{osteoblast.2}}{a numeric vector}
    \item{\code{osteoblast.3}}{a numeric vector}
    \item{\code{osteoblast.4}}{a numeric vector}
    \item{\code{osteoblast.5}}{a numeric vector}
    \item{\code{osteoblast.6}}{a numeric vector}
    \item{\code{osteoblast.7}}{a numeric vector}
    \item{\code{osteoblast.8}}{a numeric vector}
    \item{\code{osteoblast.9}}{a numeric vector}
    \item{\code{osteoblast.10}}{a numeric vector}
    \item{\code{osteoblast.11}}{a numeric vector}
    \item{\code{osteoblast.12}}{a numeric vector}
    \item{\code{osteoblast.13}}{a numeric vector}
    \item{\code{osteoblast.14}}{a numeric vector}
    \item{\code{osteoblast.15}}{a numeric vector}
    \item{\code{osteoblast.16}}{a numeric vector}
    \item{\code{osteoblast.17}}{a numeric vector}
    \item{\code{osteoblast.18}}{a numeric vector}
    \item{\code{osteoblast.19}}{a numeric vector}
    \item{\code{osteoblast.20}}{a numeric vector}
    \item{\code{osteoblast.21}}{a numeric vector}
    \item{\code{osteoblast.22}}{a numeric vector}
    \item{\code{osteoblast.23}}{a numeric vector}
    \item{\code{osteoblast.24}}{a numeric vector}
    \item{\code{osteoblast.25}}{a numeric vector}
    \item{\code{osteoblast.26}}{a numeric vector}
    \item{\code{osteoblast.27}}{a numeric vector}
    \item{\code{osteoblast.28}}{a numeric vector}
    \item{\code{osteoblast.29}}{a numeric vector}
    \item{\code{osteoblast.30}}{a numeric vector}
    \item{\code{osteoblast.31}}{a numeric vector}
    \item{\code{osteoblast.32}}{a numeric vector}
    \item{\code{osteoblast.33}}{a numeric vector}
    \item{\code{osteoblast.34}}{a numeric vector}
    \item{\code{osteoblast.35}}{a numeric vector}
    \item{\code{adipocyte}}{a numeric vector}
    \item{\code{adipocyte.1}}{a numeric vector}
    \item{\code{adipocyte.2}}{a numeric vector}
    \item{\code{adipocyte.3}}{a numeric vector}
    \item{\code{adipocyte.4}}{a numeric vector}
    \item{\code{adipocyte.5}}{a numeric vector}
    \item{\code{adipocyte.6}}{a numeric vector}
    \item{\code{adipocyte.7}}{a numeric vector}
    \item{\code{PlasmaMemory}}{a numeric vector}
    \item{\code{PlasmaMemory.1}}{a numeric vector}
    \item{\code{PlasmaMemory.2}}{a numeric vector}
    \item{\code{PlasmaMemory.3}}{a numeric vector}
  }
}
\source{
https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-3732/
https://www.ebi.ac.uk/arrayexpress/experiments/E-MEXP-3711/
https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-4152/
}
\examples{
data(addMGSM27)
## maybe str(addMGSM27) ; plot(addMGSM27) ...
}
\keyword{datasets}
