\name{cd138p}
\alias{cd138p}
\docType{data}
\title{
Purified CD138+ samples
}
\description{
CD138+ sorted cells from UAMS.  Described in Danziger et al.s 2019 paper "Identifying a High-risk Cellular Signature in the Multiple Myeloma Bone Marrow Microenvironment""
}
\usage{data("cd138p")}
\format{
  A data frame with 17480 observations on the following 426 variables.
  \describe{
    \item{\code{AMP.P0462.02.E808.U133Plus.2.cel}}{a numeric vector}
    \item{\code{AMP.P1174.05.D586.U133Plus.2.cel}}{a numeric vector}
    \item{\code{AMP.P1175.04.D359.U133Plus.2.cel}}{a numeric vector}
    \item{\code{AMP.P1465.02.E717.U133Plus.2.cel}}{a numeric vector}
    \item{\code{BI13.000755.CEL}}{a numeric vector}
    \item{\code{CL11.000085.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{CL11.000164.CEL}}{a numeric vector}
    \item{\code{CL11.000449.CEL}}{a numeric vector}
    \item{\code{CL11.000453.CEL}}{a numeric vector}
    \item{\code{CL11.000633.CEL}}{a numeric vector}
    \item{\code{CL11.000826.CEL}}{a numeric vector}
    \item{\code{CL11.000914.CEL}}{a numeric vector}
    \item{\code{CL11.000919.CEL}}{a numeric vector}
    \item{\code{CL11.000922.CEL}}{a numeric vector}
    \item{\code{CL11.000932.CEL}}{a numeric vector}
    \item{\code{CL11.001590.CEL}}{a numeric vector}
    \item{\code{CL11.001662.CEL}}{a numeric vector}
    \item{\code{CL11.001709.CEL}}{a numeric vector}
    \item{\code{CL11.001721.CEL}}{a numeric vector}
    \item{\code{CL11.001856.CEL}}{a numeric vector}
    \item{\code{CL11.001960.CEL}}{a numeric vector}
    \item{\code{CL11.001972.CEL}}{a numeric vector}
    \item{\code{CL11.002250.CEL}}{a numeric vector}
    \item{\code{CL12.000091.CEL}}{a numeric vector}
    \item{\code{CL12.000103.CEL}}{a numeric vector}
    \item{\code{CL12.000145.CEL}}{a numeric vector}
    \item{\code{CL12.000158.CEL}}{a numeric vector}
    \item{\code{CL12.000378.CEL}}{a numeric vector}
    \item{\code{CL12.000413.CEL}}{a numeric vector}
    \item{\code{CL12.000580.CEL}}{a numeric vector}
    \item{\code{CL12.000766.CEL}}{a numeric vector}
    \item{\code{CL12.000916.CEL}}{a numeric vector}
    \item{\code{CL12.001091.CEL}}{a numeric vector}
    \item{\code{CL12.001168.CEL}}{a numeric vector}
    \item{\code{CL12.001183.CEL}}{a numeric vector}
    \item{\code{CL12.001187.CEL}}{a numeric vector}
    \item{\code{CL12.001207.CEL}}{a numeric vector}
    \item{\code{CL12.001213.CEL}}{a numeric vector}
    \item{\code{CL12.001279.CEL}}{a numeric vector}
    \item{\code{CL12.001284.CEL}}{a numeric vector}
    \item{\code{CL12.001289.CEL}}{a numeric vector}
    \item{\code{CL12.001490.CEL}}{a numeric vector}
    \item{\code{CL12.001522.CEL}}{a numeric vector}
    \item{\code{CL12.001587.CEL}}{a numeric vector}
    \item{\code{CL12.001613.CEL}}{a numeric vector}
    \item{\code{CL12.001696.CEL}}{a numeric vector}
    \item{\code{CL12.001723.CEL}}{a numeric vector}
    \item{\code{CL12.001894.CEL}}{a numeric vector}
    \item{\code{CL12.002006.CEL}}{a numeric vector}
    \item{\code{CL12.002021.CEL}}{a numeric vector}
    \item{\code{CL12.002166.CEL}}{a numeric vector}
    \item{\code{CL12.002277.CEL}}{a numeric vector}
    \item{\code{CL12.002404.CEL}}{a numeric vector}
    \item{\code{CL12.002640.CEL}}{a numeric vector}
    \item{\code{CL12.002696.CEL}}{a numeric vector}
    \item{\code{CL12.003041.CEL}}{a numeric vector}
    \item{\code{CL12.003051.CEL}}{a numeric vector}
    \item{\code{CL12.003119.CEL}}{a numeric vector}
    \item{\code{CL12.003303.CEL}}{a numeric vector}
    \item{\code{CL13.000079.CEL}}{a numeric vector}
    \item{\code{CL13.000197.CEL}}{a numeric vector}
    \item{\code{CL13.000371.CEL}}{a numeric vector}
    \item{\code{CL13.001812.CEL}}{a numeric vector}
    \item{\code{CL13.001845.CEL}}{a numeric vector}
    \item{\code{CL13.001851.CEL}}{a numeric vector}
    \item{\code{CL13.001855.CEL}}{a numeric vector}
    \item{\code{CL13.002943.CEL}}{a numeric vector}
    \item{\code{P0521.02.D162.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P0990.02.C968.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P0993.00.D003.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P0995.02.D102.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P0996.04.D062.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1003.02.D033.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1010.01.D039.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1023.01.D006.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1027.02.D051.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1053.01.C795.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1063.01.C981.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1064.01.C986.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1068.01.D099.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1070.01.D245.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1077.02.D264.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1079.02.D278.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1081.02.D279.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1086.01.D205.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1088.01.D251.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1093.01.D394.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1100.01.D431.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1101.02.D459.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1112.01.D320.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1113.01.D323.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1117.00.D346.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1119.01.D487.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1120.01.C865.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1130.03.D440.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1132.03.D485.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1138.02.D405.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1141.01.D334.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1142.01.D343.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1143.01.D215.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1144.01.D361.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1145.01.D422.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1164.01.D505.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1165.01.D454.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1166.01.D468.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1167.01.D486.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1169.01.D553.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1173.01.D516.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1176.01.D600.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1177.01.D639.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1178.01.D566.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1182.01.D617.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1183.01.D417.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1184.01.D637.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1186.01.D612.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1187.01.D657.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1188.01.D658.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1189.01.D662.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1195.01.D624.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1196.01.D432.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1224.01.D803.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1225.01.D807.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1228.01.D816.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1231.01.D615.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1232.01.D646.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1239.01.D778A.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1248.01.D777.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1259.01.D865.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1263.01.D755.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1271.01.D891.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1291.02.E002.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1296.00.D980.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1305.01.E053.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1307.01.E064.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1309.01.E068.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1314.00.D963.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1320.01.E088.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1321.00.E004.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1323.01.E099.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1331.02.E169.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1336.02.E158.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1337.01.E200.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1338.01.E237.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1339.01.E247.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1341.01.E149.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1350.01.E288.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1354.01.E286.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1359.01.E293.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1363.01.E262.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1366.01.E227.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1367.01.E264.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1371.01.E336.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1375.01.E113.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1383.01.E337.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1384.01.E385.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1389.00.E033.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1390.00.E300.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1391.00.E358.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1402.01.E441.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1405.01.E440.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1407.01.E466.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1408.01.E461.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1409.01.E433.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1412.01.E507.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1424.02.E484.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1432.01.E634.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1434.01.E537.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1437.01.E510.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1441.01.E567.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1442.01.E575.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1445.01.E571.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1446.01.E504.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1448.01.E561.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1450.01.E503.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1451.01.E558.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1453.01.E515.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1458.01.E488.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1463.01.E600.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1492.01.E680.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1502.01.E768.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1503.01.E387.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1529.01.E818.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1550.01.E679.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1558.02.E821.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1571.01.E644.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1579.01.E903.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1580.01.E714.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1581.01.E800.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1583.01.E853.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1586.01.E742.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1587.01.E921.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1590.01.E719.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1591.01.E712.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1593.01.E738.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1594.01.E747.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1595.01.E769.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1596.01.E739.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1599.01.E877.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1603.01.E929.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1605.01.E926.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1606.01.E950.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1628.01.F023.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1632.07.J887.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P1634.04.F406.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1644.01.F006.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1646.01.F012.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1670.01.E997.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1691.02.F218.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1694.01.F209.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1715.01.E953.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1734.02.F294.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1735.01.F284.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1748.01.F238.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1759.01.F026.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1773.02.E795.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1774.02.F289.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1775.02.F328B.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1778.02.F157.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1779.03.F268.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1782.02.F186.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1784.02.F282.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1788.02.F195.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1789.02.F210.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1790.02.F256.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1872.01.F258.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1905.01.F607.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1908.01.F619.U133Plus.2_2.cel}}{a numeric vector}
    \item{\code{P1919.01.F644.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1939.02.F554.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1967.01.F702.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1969.04.G376.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1975.01.F655.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1978.01.F824.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1979.02.F859.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1983.01.F737.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1984.01.F673.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1986.02.G212.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1989.01.F645.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2001.01.F614.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2007.04.G168.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2011.03.F826.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2013.01.F803.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2015.01.F979.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2018.01.F935.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2022.01.G002.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2032.01.F951.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2039.02.G020.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2044.01.F848.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2045.01.F671.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2061.00.F950.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2066.01.G138.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2068.01.G221.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2079.02.G541.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2085.05.J608.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2086.02.G214.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2088.01.G257.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2097.01.G192.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2100.01.G162.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2102.01.G290.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2106.01.G114.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2107.01.G177.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2120.01.G116.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2121.01.G088.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2122.01.G363.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2129.01.G289.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2149.01.G429.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2152.01.G402.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2154.01.G332.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2155.01.G415.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2166.01.G436.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2176.01.G470.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2177.01.G483.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2187.01.G266.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2191.01.G461.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2197.01.G473.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2199.01.G277.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2204.01.G519.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2213.02.G545.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2224.01.G547.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2229.00.G558.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2232.02.G562.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2238.00.G434.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2240.01.G241.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2243.01.G580.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2252.02.G594.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2258.02.G639.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2263.02.G649.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2275.01.G699.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2278.01.G708.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2283.02.G714.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2289.04.G656.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2296.01.G603.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2304.01.G620.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2308.01.G721.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2317.01.G767.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2346.04.G859.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2355.01.G871.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2361.00.G880.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2366.01.G907.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2372.02.G968.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2408.01.H055.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2413.01.G959.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2465.02.H159.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2469.01.H247.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2471.00.H170.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2480.01.H271.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2485.01.H274.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2494.01.H102.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2516.02.H316.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2518.01.H424.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2525.03.I160.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2537.01.H308.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2548.03.H419.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2551.01.H379.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2576.01.H406.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2584.01.H459.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2587.01.H460.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2589.01.H547.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2594.02.I359.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2598.03.H778.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2602.01.H525.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2619.01.H315.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2642.01.H593.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2652.01.H605.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2693.01.H761.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2707.07.H822.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2708.01.H819.U133Plus.2_2.cel}}{a numeric vector}
    \item{\code{P2726.01.H855.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2731.03.I998.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2755.01.H933.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2757.02.I379.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2763.01.H964.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2773.01.H998.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2778.01.I007.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2788.01.I039.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2799.01.I069.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2804.01.I093.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2814.01.I099.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2823.01.I133.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2824.01.I148.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2838.01.I207.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2840.01.I199.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2841.01.I205.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2867.01.I270.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2882.01.I339.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2883.01.I338.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2898.01.I369.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2910.01.I399.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2935.01.I468.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2943.01.I458.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2949.01.I506.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2967.02.J328.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2987.01.I613.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2996.01.I657.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3011.01.I674.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3015.01.I694.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3019.01.I712.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3053.01.I788.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3066.01.I831.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3084.01.I879.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3095.01.I905.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3126.01.J001.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3141.01.J033.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3150.01.J063.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3167.01.J132.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3182.01.J196.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3184.01.J189.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3206.01.J279.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3209.01.J266.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3210.01.J263.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3212.01.J281.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3221.01.J318.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3223.01.J325.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3229.02.J528.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3231.01.J345.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3262.01.J449.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3275.01.J493.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3281.01.J502.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3288.02.J524.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3306.01.J589.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3306.02.J967.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3332.01.J660.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3365.01.J761.U133Plus.2_2.CEL}}{a numeric vector}
    \item{\code{P3386.01.J829.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3398.01.J859.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3417.01.J918.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3423.01.J938.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3433.01.J963.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3439.01.J984.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3441.01.J999.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3442.01.K001.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3447.01.K021.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3462.01.K073.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3464.01.K076.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3465.01.K067.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3469.01.K081.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3471.01.K094.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3473.01.K101.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3474.01.K102.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3476.01.K126.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3477.01.K127.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3484.01.K136.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3485.01.K137.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3529.01.K297.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3533.01.K302.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3535.01.K308.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3542.01.K329.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3558.01.K368.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3564.01.K386.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3568.01.K408.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3571.01.K404.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3583.01.K458.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3640.01.K569.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3643.01.K582.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3645.01.K579.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3650.01.K599.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3655.01.K608.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3659.01.K617.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3669.01.K706.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3674.01.K730.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3676.01.K733.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3814.01.K863.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3823.01.K884.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3841.01.K970.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3876.01.L092.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3903.01.L211.U133Plus.2.CEL}}{a numeric vector}
  }
}
\details{
Tumor burden information is available in ADAPTSdata::tumorPer
}
\source{
GEO TBD
}
\references{
Danziger et al. (2019) Identifying a High-risk Cellular Signature in the Multiple Myeloma Bone Marrow Microenvironment
}
\examples{
data(cd138p)
## maybe str(cd138p) ; plot(cd138p) ...
}
\keyword{datasets}
