\name{PBMC}
\alias{PBMC}
\docType{data}
\title{
Gene Expression from PBMCS in E-MTAB-3732
}
\description{
PBMCS have well established mixes of immune cells.
}
\usage{data("PBMC")}
\format{
  A data frame with 17481 observations on the following 182 variables.
  \describe{
    \item{\code{Sample.160.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.207.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.208.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.209.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.240.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.1138.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.1139.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.1272.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.1273.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.1274.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.1559.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.1591.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.1592.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.1620.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.2275.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.2384.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.2473.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.2518.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.2520.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.2521.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.2704.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.2937.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.2964.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.2965.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.2966.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.3008.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.3009.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.3652.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.3653.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.3927.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.4053.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.4054.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.4337.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.4351.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.4379.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.5332.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.5763.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.5801.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.6462.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.6473.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.6731.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.6732.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.6733.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.7158.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.7159.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.7248.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.7249.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.7771.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.7854.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.8012.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.8101.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.8102.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.8103.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.8244.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.8298.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.8532.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.8551.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.8552.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.8599.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.8654.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.9949.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.9950.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.9951.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.10026.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.10620.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.10621.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.10766.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.10867.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.10868.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.10869.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.11023.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.11326.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.11327.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.11372.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.11840.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.12027.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.12296.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.12299.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.12300.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.12424.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.12767.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.12806.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.13530.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.13531.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.13532.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.13638.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.13712.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.13713.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.13826.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.13827.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.13828.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.14116.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.14126.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.14150.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.14151.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.14540.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.14955.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.15081.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.15214.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.15258.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.15547.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.15548.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.15549.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.16027.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.16406.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.16482.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.16625.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.16626.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.16904.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.17010.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.17607.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.17868.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.17869.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.18013.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.18295.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.18330.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.18331.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.18394.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.18837.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.19123.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.19124.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.19217.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.19270.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.19271.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.19385.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.19386.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.19718.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.19719.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.19720.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.19721.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.19796.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.20111.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.20490.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.20491.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.20591.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.20658.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.21084.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.21186.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.21773.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.22025.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.22163.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.22164.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.22482.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.22483.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.22568.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.23350.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.23441.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.23566.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.23567.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.23844.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.23874.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.23875.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.23876.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.23913.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.23959.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.24540.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.24541.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.24794.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.24927.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.25241.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.25264.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.25265.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.25266.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.25350.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.25665.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.26062.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.26168.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.26358.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.26359.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.26360.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.26673.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.26721.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.26769.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.26770.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.27069.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.27519.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.27520.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.27564.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.27614.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.27616.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.27751.mononuclear.cell}}{a numeric vector}
    \item{\code{Sample.27800.mononuclear.cell}}{a numeric vector}
  }
}
\source{
https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-3732/
}
\examples{
data(PBMC)
## maybe str(PBMC) ; plot(PBMC) ...
}
\keyword{datasets}
